+++ src/main.go
@@ -296,10 +296,10 @@ func indexHandlerCalcDaily(w http.ResponseWriter, r *http.Request) {
 	//fmt.Fprintln(w, whole_codeRate)
 
 	// 事前にrateのシートをclear
-	clearSheet(sheetService, r, "DAILYRATE_SHEETID", "daily_rate")
+	clearSheet(sheetService, r, DAILYRATE_SHEETID, "daily_rate")
 
 	// 株価の比率順にソートしたものを書き込み
-	writeRate(sheetService, r, whole_codeRate, "DAILYRATE_SHEETID", "daily_rate")
+	writeRate(sheetService, r, whole_codeRate, DAILYRATE_SHEETID, "daily_rate")
 }
 
 func codeDateModprice(r *http.Request, resp [][]interface{}) [][]interface{} {
@@ -392,10 +392,10 @@ func indexHandler(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintln(w, whole_codeRate)
 
 	// 事前にrateのシートをclear
-	clearSheet(sheetService, r, "RATE_SHEETID", "rate")
+	clearSheet(sheetService, r, RATE_SHEETID, "rate")
 
 	// 株価の比率順にソートしたものを書き込み
-	writeRate(sheetService, r, whole_codeRate, "RATE_SHEETID", "rate")
+	writeRate(sheetService, r, whole_codeRate, RATE_SHEETID, "rate")
 }
 
 func getClientWithJson(r *http.Request) *http.Client {
@@ -423,6 +423,8 @@ var (
 	ENV                string
 	HOLIDAY_SHEETID    string
 	STOCKPRICE_SHEETID string
+	DAILYRATE_SHEETID  string
+	RATE_SHEETID       string
 )
 
 func getEnv(r *http.Request) {
@@ -439,20 +441,22 @@ func getEnv(r *http.Request) {
 	}
 	HOLIDAY_SHEETID = mustGetenv(r, "HOLIDAY_SHEETID")
 	STOCKPRICE_SHEETID = mustGetenv(r, "STOCKPRICE_SHEETID")
+	DAILYRATE_SHEETID = mustGetenv(r, "DAILYRATE_SHEETID")
+	RATE_SHEETID = mustGetenv(r, "RATE_SHEETID")
 
 }
 
-// spreadsheets clientを取得
-func getSheetClient(r *http.Request) (*sheets.Service, error) {
-	// googleAPIへのclientをリクエストから作成
-	client := getClientWithJson(r)
-	// spreadsheets clientを取得
-	srv, err := sheets.New(client)
-	if err != nil {
-		return nil, fmt.Errorf("unable to retrieve Sheets Client %v", err)
-	}
-	return srv, nil
-}
+//// spreadsheets clientを取得
+//func getSheetClient(r *http.Request) (*sheets.Service, error) {
+//	// googleAPIへのclientをリクエストから作成
+//	client := getClientWithJson(r)
+//	// spreadsheets clientを取得
+//	srv, err := sheets.New(client)
+//	if err != nil {
+//		return nil, fmt.Errorf("unable to retrieve Sheets Client %v", err)
+//	}
+//	return srv, nil
+//}
 
 func isBussinessday(srv *sheets.Service, r *http.Request) bool {
 	ctx := appengine.NewContext(r)
@@ -806,34 +810,34 @@ func writeStockprice(srv *sheets.Service, r *http.Request, code string, date str
 	}
 }
 
-func getSheetData(r *http.Request, srv *sheets.Service, sheetId string, readRange string) [][]interface{} {
-	ctx := appengine.NewContext(r)
-
-	var MaxRetries = 3
-	attempt := 0
-	for {
-		// MaxRetries を超えていたらnilを返す
-		if attempt >= MaxRetries {
-			log.Errorf(ctx, "Failed to retrieve data from sheet. attempt: %d. reached MaxRetries!", attempt)
-			return nil
-		}
-		attempt = attempt + 1
-		// stockpriceシートからデータを取得
-		resp, err := srv.Spreadsheets.Values.Get(sheetId, readRange).Do()
-		if err != nil {
-			log.Warningf(ctx, "Unable to retrieve data from sheet: %v. attempt: %d", err, attempt)
-			time.Sleep(1 * time.Second) // 1秒待つ
-			continue
-		}
-		status := resp.ServerResponse.HTTPStatusCode
-		if status != 200 {
-			log.Warningf(ctx, "HTTPstatus error: %v. attempt: %d", status, attempt)
-			time.Sleep(1 * time.Second) // 1秒待つ
-			continue
-		}
-		return resp.Values
-	}
-}
+//func getSheetData(r *http.Request, srv *sheets.Service, sheetId string, readRange string) [][]interface{} {
+//	ctx := appengine.NewContext(r)
+//
+//	var MaxRetries = 3
+//	attempt := 0
+//	for {
+//		// MaxRetries を超えていたらnilを返す
+//		if attempt >= MaxRetries {
+//			log.Errorf(ctx, "Failed to retrieve data from sheet. attempt: %d. reached MaxRetries!", attempt)
+//			return nil
+//		}
+//		attempt = attempt + 1
+//		// stockpriceシートからデータを取得
+//		resp, err := srv.Spreadsheets.Values.Get(sheetId, readRange).Do()
+//		if err != nil {
+//			log.Warningf(ctx, "Unable to retrieve data from sheet: %v. attempt: %d", err, attempt)
+//			time.Sleep(1 * time.Second) // 1秒待つ
+//			continue
+//		}
+//		status := resp.ServerResponse.HTTPStatusCode
+//		if status != 200 {
+//			log.Warningf(ctx, "HTTPstatus error: %v. attempt: %d", status, attempt)
+//			time.Sleep(1 * time.Second) // 1秒待つ
+//			continue
+//		}
+//		return resp.Values
+//	}
+//}
 
 func calcIncreaseRate(resp [][]interface{}, code string, num int, r *http.Request) ([]float64, error) {
 	ctx := appengine.NewContext(r)
@@ -884,74 +888,74 @@ func calcIncreaseRate(resp [][]interface{}, code string, num int, r *http.Reques
 	return rate, nil
 }
 
-func clearSheet(srv *sheets.Service, r *http.Request, sid string, sname string) {
-	ctx := appengine.NewContext(r)
+//func clearSheet(srv *sheets.Service, r *http.Request, sid string, sname string) {
+//	ctx := appengine.NewContext(r)
+//
+//	sheetId := ""
+//	// sheetIdを環境変数から読み込む
+//	if v := os.Getenv(sid); v != "" {
+//		sheetId = v
+//	} else {
+//		log.Errorf(ctx, "Failed to get price rate sheetId. '%v'", v)
+//		os.Exit(0)
+//	}
+//	writeRange := sname
+//
+//	// clear stockprice rate spreadsheet:
+//	resp, err := srv.Spreadsheets.Values.Clear(sheetId, writeRange, &sheets.ClearValuesRequest{}).Do()
+//	if err != nil {
+//		log.Errorf(ctx, "Unable to clear value. %v", err)
+//		return
+//	}
+//	status := resp.ServerResponse.HTTPStatusCode
+//	if status != 200 {
+//		log.Errorf(ctx, "HTTPstatus error. %v", status)
+//		return
+//	}
+//}
 
-	sheetId := ""
-	// sheetIdを環境変数から読み込む
-	if v := os.Getenv(sid); v != "" {
-		sheetId = v
-	} else {
-		log.Errorf(ctx, "Failed to get price rate sheetId. '%v'", v)
-		os.Exit(0)
-	}
-	writeRange := sname
-
-	// clear stockprice rate spreadsheet:
-	resp, err := srv.Spreadsheets.Values.Clear(sheetId, writeRange, &sheets.ClearValuesRequest{}).Do()
-	if err != nil {
-		log.Errorf(ctx, "Unable to clear value. %v", err)
-		return
-	}
-	status := resp.ServerResponse.HTTPStatusCode
-	if status != 200 {
-		log.Errorf(ctx, "HTTPstatus error. %v", status)
-		return
-	}
-}
-
-func writeRate(srv *sheets.Service, r *http.Request, rate []codeRate, sid string, sname string) {
-	ctx := appengine.NewContext(r)
-
-	sheetId := ""
-	// sheetIdを環境変数から読み込む
-	if v := os.Getenv(sid); v != "" {
-		sheetId = v
-	} else {
-		log.Errorf(ctx, "Failed to get price rate sheetId. '%v'", v)
-		os.Exit(0)
-	}
-	writeRange := sname
-
-	// spreadsheetに書き込み対象の行列を作成
-	matrix := make([][]interface{}, len(rate))
-	// 株価の比率順にソートしたものを書き込み
-	//for i, r := range rate {
-	//matrix[i] = []interface{}{r.Code, r.Rate[0], r.Rate[1], r.Rate[2], r.Rate[3], r.Rate[4], r.Rate[5]}
-	//}
-	for _, r := range rate {
-		m := make([]interface{}, 0)
-		m = append(m, r.Code)
-		// Rateの個数だけ書き込み
-		for i := 0; i < len(r.Rate); i++ {
-			m = append(m, r.Rate[i])
-		}
-		matrix = append(matrix, m)
-	}
-
-	valueRange := &sheets.ValueRange{
-		MajorDimension: "ROWS",
-		Values:         matrix,
-	}
-	// Write stockprice rate spreadsheet:
-	resp, err := srv.Spreadsheets.Values.Append(sheetId, writeRange, valueRange).ValueInputOption("USER_ENTERED").InsertDataOption("INSERT_ROWS").Do()
-	if err != nil {
-		log.Errorf(ctx, "Unable to write value. %v", err)
-		return
-	}
-	status := resp.ServerResponse.HTTPStatusCode
-	if status != 200 {
-		log.Errorf(ctx, "HTTPstatus error. %v", status)
-		return
-	}
-}
+//func writeRate(srv *sheets.Service, r *http.Request, rate []codeRate, sid string, sname string) {
+//	ctx := appengine.NewContext(r)
+//
+//	sheetId := ""
+//	// sheetIdを環境変数から読み込む
+//	if v := os.Getenv(sid); v != "" {
+//		sheetId = v
+//	} else {
+//		log.Errorf(ctx, "Failed to get price rate sheetId. '%v'", v)
+//		os.Exit(0)
+//	}
+//	writeRange := sname
+//
+//	// spreadsheetに書き込み対象の行列を作成
+//	matrix := make([][]interface{}, len(rate))
+//	// 株価の比率順にソートしたものを書き込み
+//	//for i, r := range rate {
+//	//matrix[i] = []interface{}{r.Code, r.Rate[0], r.Rate[1], r.Rate[2], r.Rate[3], r.Rate[4], r.Rate[5]}
+//	//}
+//	for _, r := range rate {
+//		m := make([]interface{}, 0)
+//		m = append(m, r.Code)
+//		// Rateの個数だけ書き込み
+//		for i := 0; i < len(r.Rate); i++ {
+//			m = append(m, r.Rate[i])
+//		}
+//		matrix = append(matrix, m)
+//	}
+//
+//	valueRange := &sheets.ValueRange{
+//		MajorDimension: "ROWS",
+//		Values:         matrix,
+//	}
+//	// Write stockprice rate spreadsheet:
+//	resp, err := srv.Spreadsheets.Values.Append(sheetId, writeRange, valueRange).ValueInputOption("USER_ENTERED").InsertDataOption("INSERT_ROWS").Do()
+//	if err != nil {
+//		log.Errorf(ctx, "Unable to write value. %v", err)
+//		return
+//	}
+//	status := resp.ServerResponse.HTTPStatusCode
+//	if status != 200 {
+//		log.Errorf(ctx, "HTTPstatus error. %v", status)
+//		return
+//	}
+//}
